Decoupling Open Search Related Extension
