+++ Client/Frontend/Browser/Favorites/FavoritesViewController.swift
@@ -171,10 +171,7 @@ extension FavoritesViewController: KeyboardHelperDelegate {
     func keyboardHelper(_ keyboardHelper: KeyboardHelper, keyboardWillShowWithState state: KeyboardState) {
         updateKeyboardInset(state)
     }
-    
-    func keyboardHelper(_ keyboardHelper: KeyboardHelper, keyboardDidShowWithState state: KeyboardState) {
-    }
-    
+
     func keyboardHelper(_ keyboardHelper: KeyboardHelper, keyboardWillHideWithState state: KeyboardState) {
         updateKeyboardInset(state)
     }
